reorganize util
